+++ lib/screens/auth_jail.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:peercoin/providers/appsettings.dart';
-import 'package:peercoin/providers/encryptedbox.dart';
+import 'package:peercoin/providers/app_settings.dart';
+import 'package:peercoin/providers/encrypted_box.dart';
 import 'package:peercoin/tools/app_localizations.dart';
 import 'package:peercoin/tools/app_routes.dart';
 import 'package:peercoin/tools/auth.dart';