+++ lib/screens/server_settings.dart
@@ -1,7 +1,7 @@
 import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter/material.dart';
 import 'package:peercoin/models/server.dart';
-import 'package:peercoin/providers/electrumconnection.dart';
+import 'package:peercoin/providers/electrum_connection.dart';
 import 'package:peercoin/providers/servers.dart';
 import 'package:peercoin/tools/app_localizations.dart';
 import 'package:peercoin/tools/app_routes.dart';