+++ lib/screens/setup/setup.dart
@@ -1,7 +1,7 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:peercoin/tools/app_localizations.dart';
-import 'package:peercoin/providers/activewallets.dart';
+import 'package:peercoin/providers/active_wallets.dart';
 import 'package:peercoin/tools/app_routes.dart';
 import 'package:peercoin/widgets/buttons.dart';
 import 'package:peercoin/widgets/setup_progress.dart';