+++ lib/screens/wallet/import_wif.dart
@@ -1,13 +1,13 @@
 import 'package:coinslib/coinslib.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:peercoin/models/availablecoins.dart';
+import 'package:peercoin/models/available_coins.dart';
 import 'package:peercoin/models/coin.dart';
-import 'package:peercoin/providers/activewallets.dart';
-import 'package:peercoin/providers/electrumconnection.dart';
+import 'package:peercoin/providers/active_wallets.dart';
+import 'package:peercoin/providers/electrum_connection.dart';
 import 'package:peercoin/tools/app_localizations.dart';
 import 'package:peercoin/tools/app_routes.dart';
-import 'package:peercoin/tools/backgroundsync.dart';
+import 'package:peercoin/tools/background_sync.dart';
 import 'package:peercoin/widgets/buttons.dart';
 import 'package:provider/provider.dart';