+++ lib/screens/wallet/transaction_details.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:peercoin/tools/app_localizations.dart';
-import 'package:peercoin/models/availablecoins.dart';
-import 'package:peercoin/models/coinwallet.dart';
-import 'package:peercoin/models/wallettransaction.dart';
+import 'package:peercoin/models/available_coins.dart';
+import 'package:peercoin/models/coin_wallet.dart';
+import 'package:peercoin/models/wallet_transaction.dart';
 import 'package:peercoin/widgets/buttons.dart';
 import 'package:url_launcher/url_launcher.dart';