+++ lib/widgets/wallet/new_wallet.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:peercoin/providers/appsettings.dart';
-import 'package:peercoin/providers/unencryptedOptions.dart';
+import 'package:peercoin/providers/app_settings.dart';
+import 'package:peercoin/providers/unencrypted_options.dart';
 import 'package:peercoin/tools/app_localizations.dart';
-import 'package:peercoin/models/availablecoins.dart';
+import 'package:peercoin/models/available_coins.dart';
 import 'package:peercoin/models/coin.dart';
-import 'package:peercoin/providers/activewallets.dart';
+import 'package:peercoin/providers/active_wallets.dart';
 import 'package:peercoin/tools/app_routes.dart';
 import 'package:peercoin/tools/auth.dart';
 import 'package:provider/provider.dart';