+++ test_driver/key_new.dart
@@ -64,6 +64,7 @@ void main() {
             <String>['shell', 'input', 'keyevent', 'KEYCODE_BACK'],
             runInShell: true,
           );
+          await driver.tap(find.text('Okay'));
           expect(
             await driver.getText(find.byValueKey('noActiveWallets')),
             'You have no active wallets',